Fixed mul instruction
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@3041 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/arch/mips/emu/ContextIsa.cc
@@ -1157,7 +1157,7 @@ void Context::ExecuteInst_MUL()
 	unsigned int rd = inst.getBytes()->standard.rd;
 
 	// Perform operation
-	long int temp = (long int)regs.getGPR(rs) * (long int)regs.getGPR(rt);
+	long long int temp = (int)regs.getGPR(rs) * (int)regs.getGPR(rt);
 	regs.setGPR(rd, temp);
 }
